requirements.txt
flask== 1.0.2
pyyaml==3.13
python-dateutil==2.8.0
SQLAlchemy==1.3.1
flask_sqlalchemy==2.3.2
pytest==4.0.1
boto3==1.9.82
pandas==0.23.4
scikit-learn==0.20.3
xgboost==0.82